Move SEV-SNP attestation report structure to own crate.
Motivation is to make it easier to use from other crates by reducing transitive dependencies.

b/322156740
