+++ oak_containers_kernel/configs/6.1.33/minimal.config
@@ -2,9 +2,9 @@
 # Automatically generated file; DO NOT EDIT.
 # Linux/x86 6.1.33 Kernel Configuration
 #
-CONFIG_CC_VERSION_TEXT="gcc (GCC) 12.3.0"
+CONFIG_CC_VERSION_TEXT="gcc (GCC) 13.2.0"
 CONFIG_CC_IS_GCC=y
-CONFIG_GCC_VERSION=120300
+CONFIG_GCC_VERSION=130200
 CONFIG_CLANG_VERSION=0
 CONFIG_AS_IS_GNU=y
 CONFIG_AS_VERSION=24000
