+++ stage0/Cargo.toml
@@ -18,6 +18,7 @@ oak_dice = { workspace = true }
 oak_stage0_dice = { workspace = true }
 oak_linux_boot_params = { path = "../linux_boot_params" }
 oak_sev_guest = { workspace = true, features = ["rust-crypto"] }
+oak_sev_snp_attestation_report = { workspace = true }
 p256 = { version = "*", default-features = false, features = ["ecdsa"] }
 rand_core = { version = "*", default-features = false, features = [
   "getrandom",
