+++ stage0_dice/Cargo.toml
@@ -10,7 +10,7 @@ coset = { version = "*", default-features = false }
 hkdf = { version = "*", default-features = false }
 hex = { version = "*", default-features = false, features = ["alloc"] }
 oak_dice = { workspace = true }
-oak_sev_guest = { workspace = true, features = ["rust-crypto"] }
+oak_sev_snp_attestation_report = { workspace = true }
 p256 = { version = "*", default-features = false, features = ["ecdsa"] }
 rand_core = { version = "*", default-features = false, features = [
   "getrandom",
